Move ValueIterator out of TLVFactory trait
